# Improved performance to 14 seconds (-6 seconds)