```
hyperfine --warmup 1 --runs 5 --export-json me-timing.json ./calculate_average_bytesfellow.sh
Benchmark 1: ./calculate_average_bytesfellow.sh
  Time (mean ± σ):     14.032 s ±  0.460 s    [User: 101.771 s, System: 4.239 s]
  Range (min … max):   13.609 s … 14.556 s    5 runs
```


macbook pro m1 10 cores.
baseline runs in 189 seconds on the same generated input file.